Task-2
This project aims to build predictive models for determining the quality of wine based on its chemical characteristics. By leveraging machine learning techniques, we can uncover patterns in the data and develop robust models to forecast wine quality, which has practical applications in the viticulture industry.
The dataset used in this project contains various chemical attributes of wines, such as density, acidity, and alcohol content, along with a quality rating for each wine sample. The goal is to train classification models to accurately predict the wine quality based on these input features.
In this project these technology are used :- 
Python
Pandas
NumPy
Scikit-learn (for machine learning models)
Seaborn and Matplotlib (for data visualization)