Task-4
This project focuses on exploring and analyzing the Google Play Store dataset to uncover insights into the dynamics of the Android app market. By leveraging data cleaning, categorization, and visualization techniques, the goal is to provide a thorough understanding of the market landscape, app trends, and user sentiments.
In this project these technology are used :- 
Python
Pandas (for data manipulation)
NumPy (for numerical operations)
Matplotlib and Seaborn (for data visualization)
Plotly (for interactive visualizations)
NLTK or other sentiment analysis libraries (for review analysis).